Fix select2 not enabling on fields added later with cocoon. I am not entirely sure what fixed it, and I have a vague feeling somewhere there is a duplicate jQuery import
