+++ app/javascript/shared.js
@@ -1,9 +1,15 @@
+import jQuery from 'jquery';
+
+global.$ = global.jQuery = jQuery;
+
+
 import "./sweetalert"
 
 import 'jquery-slimscroll'
 
 import './src/shared/konami_code'
 import './src/shared/md_editor'
+import './src/shared/select2'
 
 // Load all the stimulus controllers
 import "./controllers"
@@ -12,4 +18,3 @@ require("@rails/activestorage").start()
 
 // Requires jQuery. There are vanilla js packages, but not as frequently maintained or downloaded.
 require("@nathanvda/cocoon")
-
